Clean up CMR files
