+++ packages/manager/src/features/Firewalls/FirewallLanding/FirewallActionMenu.tsx
@@ -1,13 +1,25 @@
 import { FirewallStatus } from '@linode/api-v4/lib/firewalls';
 import * as React from 'react';
-import { useHistory } from 'react-router-dom';
+import {
+  makeStyles,
+  Theme,
+  useTheme,
+  useMediaQuery
+} from 'src/components/core/styles';
+import ActionMenu, { Action } from 'src/components/ActionMenu_CMR';
+import InlineMenuAction from 'src/components/InlineMenuAction';
 
-import ActionMenu, { Action } from 'src/components/ActionMenu/ActionMenu';
+const useStyles = makeStyles(() => ({
+  root: {
+    display: 'flex'
+  }
+}));
 
 export interface ActionHandlers {
   triggerEnableFirewall: (firewallID: number, firewallLabel: string) => void;
   triggerDisableFirewall: (firewallID: number, firewallLabel: string) => void;
   triggerDeleteFirewall: (firewallID: number, firewallLabel: string) => void;
+  [index: string]: any;
 }
 
 interface Props extends ActionHandlers {
@@ -19,6 +31,10 @@ interface Props extends ActionHandlers {
 type CombinedProps = Props;
 
 const FirewallActionMenu: React.FC<CombinedProps> = props => {
+  const classes = useStyles();
+  const theme = useTheme<Theme>();
+  const matchesSmDown = useMediaQuery(theme.breakpoints.down('sm'));
+
   const {
     firewallID,
     firewallLabel,
@@ -27,36 +43,45 @@ const FirewallActionMenu: React.FC<CombinedProps> = props => {
     triggerDisableFirewall,
     triggerDeleteFirewall
   } = props;
-  const history = useHistory();
 
-  const createActions = () => {
-    return (closeMenu: Function): Action[] => [
+  const inlineActions = [
+    {
+      actionText:
+        firewallStatus === ('enabled' as FirewallStatus) ? 'Disable' : 'Enable',
+      onClick: () => {
+        handleEnableDisable();
+      }
+    },
+    {
+      actionText: 'Delete',
+      onClick: () => {
+        triggerDeleteFirewall(firewallID, firewallLabel);
+      }
+    }
+  ];
+
+  const handleEnableDisable = () => {
+    const request = () =>
+      firewallStatus === 'disabled'
+        ? triggerEnableFirewall(firewallID, firewallLabel)
+        : triggerDisableFirewall(firewallID, firewallLabel);
+    request();
+  };
+
+  const createActions = () => (): Action[] => {
+    return [
       {
-        title: firewallStatus === 'disabled' ? 'Enable' : 'Disable',
+        title:
+          firewallStatus === ('enabled' as FirewallStatus)
+            ? 'Disable'
+            : 'Enable',
         onClick: () => {
-          const request = () =>
-            firewallStatus === 'disabled'
-              ? triggerEnableFirewall(firewallID, firewallLabel)
-              : triggerDisableFirewall(firewallID, firewallLabel);
-
-          request();
-
-          closeMenu();
-        }
-      },
-      {
-        title: 'Edit',
-        onClick: (e: React.MouseEvent<HTMLElement>) => {
-          closeMenu();
-          history.push(`/firewalls/${firewallID}/rules`);
-          e.preventDefault();
-          e.stopPropagation();
+          handleEnableDisable();
         }
       },
       {
         title: 'Delete',
         onClick: () => {
-          closeMenu();
           triggerDeleteFirewall(firewallID, firewallLabel);
         }
       }
@@ -64,10 +89,24 @@ const FirewallActionMenu: React.FC<CombinedProps> = props => {
   };
 
   return (
-    <ActionMenu
-      createActions={createActions()}
-      ariaLabel={`Action menu for Firewall ${props.firewallLabel}`}
-    />
+    <div className={classes.root}>
+      {!matchesSmDown &&
+        inlineActions.map(action => {
+          return (
+            <InlineMenuAction
+              key={action.actionText}
+              actionText={action.actionText}
+              onClick={action.onClick}
+            />
+          );
+        })}
+      {matchesSmDown && (
+        <ActionMenu
+          createActions={createActions()}
+          ariaLabel={`Action menu for Firewall ${props.firewallLabel}`}
+        />
+      )}
+    </div>
   );
 };
