+++ packages/manager/src/features/Firewalls/index.tsx
@@ -2,12 +2,8 @@ import * as React from 'react';
 import { Route, RouteComponentProps, Switch } from 'react-router-dom';
 import { DocumentTitleSegment } from 'src/components/DocumentTitle';
 import SuspenseLoader from 'src/components/SuspenseLoader';
-import useFlags from 'src/hooks/useFlags';
 
 const FirewallLanding = React.lazy(() => import('./FirewallLanding'));
-const FirewallLanding_CMR = React.lazy(() =>
-  import('./FirewallLanding/FirewallLanding_CMR')
-);
 const FirewallDetail = React.lazy(() => import('./FirewallDetail'));
 
 type Props = RouteComponentProps<{}>;
@@ -19,18 +15,12 @@ const Firewall: React.FC<CombinedProps> = props => {
     match: { path }
   } = props;
 
-  const flags = useFlags();
-
   return (
     <React.Suspense fallback={<SuspenseLoader />}>
       <React.Fragment>
         <DocumentTitleSegment segment="Firewalls" />
         <Switch>
-          <Route
-            exact
-            path={`${path}(/create)?`}
-            component={flags.cmr ? FirewallLanding_CMR : FirewallLanding}
-          />
+          <Route exact path={`${path}(/create)?`} component={FirewallLanding} />
           <Route path={`${path}/:id`} component={FirewallDetail} />
           <Route component={FirewallLanding} />
         </Switch>
